requirements.txt
arrow==0.15.5
certifi==2019.11.28
chardet==3.0.4
idna==2.9
pyaml==19.12.0
python-dateutil==2.8.1
PyYAML==5.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
xlrd==1.2.0
XlsxWriter==1.2.8
xlutils==2.0.0
xlwt==1.3.0
